Add files via upload
There is now an actual Sprite class and ObjectLayer class! I haven't changed the name of ObjectLayer to SpriteLayer just because we have a member of that class called SpriteLayer, but that should also be done. Arrow keys will move the sprites, just to demonstrate that the sprites are not static.