With the current set up, I fixed an issue with undoRandom Move which undid a movement, regardless of whether it was adjusted by the boarders of the level.
Currently the issue seems to be that down and left may be seeing collision one pixel further than it should, which ends up with it being placed on a stack, allowing another object to pass that gap (by going right or up), and then the overlap isn't registered potentially.  This fix will take some time, simply because of it's complexity.