In Main: Added a check to see if the velocity would be appropriately adjusted on the key being lifted up. This is because my check isn't quite perfect and will recheck a sprite and adjust it's velocity while it's in that position, instead of just whenever the velocity is appropriate to change.
In Global: Added a singular array of 8 integers, used to help me pause when dealing with rare occurances just once.

in Sprite.h: Figured out jostling so that when we would teleport on one axis for a diagonal movement, after the overlap is figured out (if any) it will then check it's foreseeable new position for the first collision it'd end up with, and move back accordingly. This is entirely dependent on hitting at least one sprite on the 'edges' of the diagonal movement. Thus a bug occured when moving around the edge of the map. To fix this I'm making it so the velocity at the edge of the map is fixed so that you cannot go diagonally around the corners. The button presses are still occuring, or are down, it's just the velocities are changed. Thus it shouldn't be an issue for graphical displays or manuvers later.