This bug is going to be a bit of a pain. It seems that because I don't start with the jostling, it doesn't read the collision on the LR corner as a possiblity, and thus it is ignored. Usually, that'd be fine, but since we collided with something on the other axis, we can be reasonably certain that we will end up sliding downwards. Thus I'll have to make a check to figure out what sprites are on the LR corner, and figure out the appropriate ptr based on overlap. -_-. This will also have to happen for all 7 other instances where one axis could be triggered, necessitating the jostling in the future. To trigger that jostle we need at least one ptr in the right direction.
The other approach I could try, is just  to brute force check all four points of interest all the time. Although computationally wasteful. It may not work either but it's a hail merry.

I'll try the second approach first, and see how it operates. If it's an issue I'll revert to this version and try the first.