Will continue to work after a short break. Found some bug's regarding the relative position formulas, just got a bit to many copy pastes without thinking. Will fix those tonight, and then see if that does the trick.
Sliding worked for a bit, but then I changed the diagonal check (Since there was a slight error, '=' instead of '==', for some reason didn't trigger the error warning), as well as passing a reference of TempStackable so that we can continually see the information for each collision as they happen.
The jostling logic and the overlap are the two runaway while's that happen. The checkoverlap will be dealt with shortly, and is in large part because I don't have a stack handle yet, and the jostling logic I believe is from these formulas. These formulas are what cause the overlap, so it should finish off both bugs.